fix: move parts to sections and create attachment as a class
