+++ app/services/pdf/beratungshilfe/beratungshilfe.pdf.ts
@@ -1,5 +1,5 @@
 import { CheckboxValue } from "~/components/inputs/Checkbox";
-import { BeratungshilfeFormularContext } from "~/models/flows/beratungshilfeFormular";
+import type { BeratungshilfeFormularContext } from "~/models/flows/beratungshilfeFormular";
 
 export const staatlicheLeistungMapping = {
   grundsicherung: "Grundsicherung",
@@ -59,68 +59,6 @@ export const getOccupationDetails = (
   }
 
   return description.filter((value) => value).join(", ");
-};
-
-export const newPageHint = "Bitte im Anhang prüfen";
-
-export const isANewAttachmentPageNeeded = (
-  context: BeratungshilfeFormularContext,
-): Attachment => {
-  const descriptions = [];
-
-  if (context.bereich) {
-    // TODO move to another function and use strapi as a source
-    const bereichMapping = {
-      authorities: "Behörden",
-      living: "Wohnen",
-      work: "Arbeit",
-      separation: "Trennung & Unterhalt",
-      trade: "Handel & Verträge",
-      debt: "Schulden & Forderungen",
-      inheritance: "Erben",
-      criminalProcedure: "Strafverfahren",
-      other: "Sonstiges",
-    };
-
-    descriptions.push({
-      title: "Thema des Rechtsproblems:",
-      text: bereichMapping[context.bereich],
-    });
-  }
-
-  if (context.beschreibung) {
-    descriptions.push({
-      title: "Beschreibung Angelegenheit:",
-      text: context.beschreibung,
-    });
-  }
-
-  if (context.eigeninitiativeBeschreibung) {
-    descriptions.push({
-      title: "Eigenbemühungen:",
-      text: context.eigeninitiativeBeschreibung,
-    });
-  } else if (context.keineEigeninitiativeBeschreibung) {
-    descriptions.push({
-      title: "Keine Eigenbemühung, weil:",
-      text: context.keineEigeninitiativeBeschreibung,
-    });
-  }
-
-  if (context.sonstiges) {
-    descriptions.push({ title: "Weitere Anmerkung:", text: context.sonstiges });
-  }
-
-  return {
-    shouldCreateNewPage:
-      descriptions.map((x) => x.title + x.text).join(" ").length > 255,
-    descriptions,
-  };
-};
-
-export type Attachment = {
-  shouldCreateNewPage: boolean;
-  descriptions: { title: string; text: string }[];
 };
 
 export function getSelectedOptions(
