+++ app/services/pdf/beratungshilfe/sections/C_einkommen.ts
@@ -1,0 +1,18 @@
+import type { BeratungshilfeFormularContext } from "~/models/flows/beratungshilfeFormular";
+import type { BeratungshilfePDF } from "../beratungshilfe.generated";
+
+export function fillEinkommen(
+  pdfFields: BeratungshilfePDF,
+  context: BeratungshilfeFormularContext,
+) {
+  pdfFields.c2Einkuenftenetto.value = context.einkommen;
+
+  if (
+    context.partnerschaft === "yes" &&
+    context.zusammenleben === "yes" &&
+    context.partnerEinkommen === "yes"
+  ) {
+    pdfFields.c3EinkuenftePartner.value = true;
+    pdfFields.c4EinkuenftePartnernetto.value = context.partnerEinkommenSumme;
+  }
+}
