+++ app/services/pdf/beratungshilfe/sections/footer.ts
@@ -1,0 +1,19 @@
+import type { BeratungshilfeFormularContext } from "~/models/flows/beratungshilfeFormular";
+import type { BeratungshilfePDF } from "../beratungshilfe.generated";
+
+export function fillFooter(
+  pdfFields: BeratungshilfePDF,
+  context: BeratungshilfeFormularContext,
+) {
+  const address = [
+    context.anwaltName ?? "",
+    context.anwaltStrasseUndHausnummer ?? "",
+    context.anwaltPlz ?? "",
+    context.anwaltOrt ?? "",
+  ]
+    .filter((entry) => entry)
+    .join(", ");
+
+  pdfFields.beratungsperson.value = address;
+  pdfFields.datumBeratung.value = context.beratungStattgefundenDatum;
+}
